+++ changelog.md
@@ -4,6 +4,7 @@ __1.14.0__
 
 * PatchMatch algorithm for image editing, inpainting and completion, in progress
 * FILTER.Util.Image.Selection class
+* handle metaData serialization in CompositeFilter
 
 __1.13.0__